automation/requirements.txt
cryptography==43.0.1
jwt==1.3.1
pandoc-mustache==0.1.0
Pillow==10.4.0
pyyaml==6.0.2
requests==2.32.3
